Add inflate/deflate semantics
Also an example inventory

Trying to get equip/dequip working was quite tricky, but this makes quite a bit of progress!

The idea is that an inflated is in the scene (=gameObject), while a deflated is outside of the scene (=scriptableobject). Each references the other, so you can have prefabs-referencing-prefabs, which is otherwise pretty hard to do safely (barring Loadables which still require unique names).

Not sure if there's an easy way to automate creating the two halves, but they do serve different purposes, so... meh?